Move gaussian drawing into the `ShapeRenderer`
